server/requirements.txt
flask
pdfplumber
google-generativeai
faiss-cpu
numpy
python-dotenv
sentence-transformers
